Add files via upload
Moving the iceberg basal melting from the surface layer (evaporation-precipitation [kg/m2/s]) to the layer that matches the keel depth. The impact of basal melting is represented as salinity and temperature tendencies per second. These tendency fields are updated by each iceberg and also exchanged. For diagnostic output, the global summation DDPDD is used. It is important to note that in the middle of each time step, the tendencies in the TS array are set to zero. We add the temperature and salinity tendencies associated with iceberg basal melt after this reset.